Fix integer conversion of string arguments in ProgressiveHedging

+++ pyomo/pysp/ph.py
@@ -1941,10 +1941,10 @@ class ProgressiveHedging(_PHBase):
         # and in that order.
         def convert_value_string_to_number(s):
             try:
-                return float(s)
+                return int(s)
             except ValueError:
                 try:
-                    return int(s)
+                    return float(s)
                 except ValueError:
                     return s
